// src/api/model/location.ts
/**
 * Generated by orval v6.25.0 🍺
 * Do not edit manually.
 * Ticketerie API
 * OpenAPI spec version: 1.0.0
 */

export interface Location {
  address: string;
  city: string;
  commonSports: number[];
  country: string;
  createdAt: string;
  id: number;
  latitude: number;
  longitude: number;
  name: string;
  state: string;
  updatedAt: string | null;
  zipcode: number;
}